Reorganize unittests
